backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
gunicorn
python-dotenv
cryptography